requirements.txt
ansi2html==1.8.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bioformats==0.1.15
bleach==5.0.0
boto3==1.23.0
botocore==1.26.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
dash==2.5.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
easygui==0.98.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
Flask-Compress==1.12
fonttools==4.33.3
future==0.18.2
idna==3.3
imageio==2.19.2
importlib-metadata==4.12.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-dash==0.4.2
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
matplotlib-scalebar==0.8.1
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.11
numpy==1.22.3
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==22.1
plotly==5.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyfaidx==0.6.4
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-bioformats==4.0.5
python-dateutil==2.8.2
python-javabridge==4.0.3
pytz==2022.1
PyVCF==0.4.3
PyWavelets==1.3.0
pyzmq==23.0.0
requests==2.28.1
retrying==1.3.3
s3transfer==0.5.2
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==61.2.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.13.2
TCSPC-Tools==0.1
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tiffwrapper==0.1
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
wheel==0.37.1
zipp==3.8.0
